pyproject.toml
[tool.poetry]
name = "guzo"
version = "0.1.0"
description = "A ridesharing and charter platform for Ethiopia"
authors = ["Guzo Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
motor = "^3.6.0"
beanie = "^1.26.0"
python-multipart = "^0.0.12"
jinja2 = "^3.1.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"  # Pin to 4.0.x for passlib compatibility
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.6.0"
sse-starlette = "^2.1.0"
redis = "^5.0.0"
httpx = "^0.27.0"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

